Linter: Rename `Skill` to `Skillfulness` since we need the word `skills` elswhere.
Comment: This is because in german `Geschicklichkeit` can be translated to `Skill` as well as `Fertigkeiten`. So we choose `Skillfulness` for `Geschicklichkeit` and `Skills` for `Fertigkeiten`. We also considered `Ability` for `Fertigkeiten` but this would more mean `Fähigkeiten` which has a similar, but kinda different meaning;

Comment 2: > "Fähigkeit" refers to an inherent ability or talent, something you are naturally good at. "Fertigkeit," on the other hand, is a skill that you have developed through practice and learning. Essentially, "Fähigkeit" is about potential, while "Fertigkeit" is about acquired proficiency.;